BOOKING-172: Moved enum
