+++ src/Enum/NotificationTypeEnum.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service;
+namespace App\Enum;
 
 enum NotificationTypeEnum
 {
