+++ src/EventListener/FailedMessageEventListener.php
@@ -2,12 +2,12 @@
 
 namespace App\EventListener;
 
+use App\Enum\NotificationTypeEnum;
 use App\Message\CreateBookingMessage;
 use App\Message\SendBookingNotificationMessage;
 use App\Message\WebformSubmitMessage;
 use App\Repository\Main\AAKResourceRepository;
 use App\Service\NotificationServiceInterface;
-use App\Service\NotificationTypeEnum;
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
 use Symfony\Component\Messenger\Event\WorkerMessageFailedEvent;
